chore: replace peer4 library with source
